+++ lib/broadside/deploy/ecs_deploy.rb
@@ -14,12 +14,11 @@ module Broadside
 
     def initialize(opts)
       super(opts)
-      config.ecs.verify(:cluster, :poll_frequency)
     end
 
     def deploy
       super do
-        unless EcsManager.service_exists?(config.ecs.cluster, family)
+        unless EcsManager.service_exists?(@deploy_config.cluster, family)
           exception "No service for #{family}! Please bootstrap or manually configure the service."
         end
         unless EcsManager.get_latest_task_definition_arn(family)
@@ -66,7 +65,7 @@ module Broadside
         run_bootstrap_commands
       end
 
-      if EcsManager.service_exists?(config.ecs.cluster, family)
+      if EcsManager.service_exists?(@deploy_config.cluster, family)
         info("Service for #{family} already exists.")
       else
         unless @deploy_config.service_config
@@ -74,7 +73,7 @@ module Broadside
         end
 
         info "Service '#{family}' doesn't exist, creating..."
-        EcsManager.create_service(config.ecs.cluster, family, @deploy_config.service_config)
+        EcsManager.create_service(@deploy_config.cluster, family, @deploy_config.service_config)
       end
     end
 
@@ -133,7 +132,7 @@ module Broadside
 
     def status
       super do
-        ips = EcsManager.get_running_instance_ips(config.ecs.cluster, family)
+        ips = EcsManager.get_running_instance_ips(@deploy_config.cluster, family)
         info "\n---------------",
           "\nDeployed task definition information:\n",
           Rainbow(PP.pp(EcsManager.get_latest_task_definition(family), '')).blue,
@@ -177,7 +176,7 @@ module Broadside
     private
 
     def get_running_instance_ip
-      EcsManager.get_running_instance_ips(config.ecs.cluster, family).fetch(@deploy_config.instance)
+      EcsManager.get_running_instance_ips(@deploy_config.cluster, family).fetch(@deploy_config.instance)
     end
 
     # Creates a new task revision using current directory's env vars, provided tag, and configured options.
@@ -206,7 +205,7 @@ module Broadside
       debug "Updating #{family} with scale=#{@deploy_config.scale} using task #{task_definition_arn}..."
 
       update_service_response = EcsManager.ecs.update_service({
-        cluster: config.ecs.cluster,
+        cluster: @deploy_config.cluster,
         desired_count: @deploy_config.scale,
         service: family,
         task_definition: task_definition_arn
@@ -216,9 +215,9 @@ module Broadside
         exception('Failed to update service during deploy.', update_service_response.pretty_inspect)
       end
 
-      EcsManager.ecs.wait_until(:services_stable, { cluster: config.ecs.cluster, services: [family] }) do |w|
-        w.max_attempts = @deploy_config.timeout ? @deploy_config.timeout / config.ecs.poll_frequency : nil
-        w.delay = config.ecs.poll_frequency
+      EcsManager.ecs.wait_until(:services_stable, { cluster: @deploy_config.cluster, services: [family] }) do |w|
+        w.max_attempts = @deploy_config.timeout ? @deploy_config.timeout / @deploy_config.poll_frequency : nil
+        w.delay = @deploy_config.poll_frequency
         seen_event = nil
 
         w.before_wait do |attempt, response|
@@ -234,7 +233,7 @@ module Broadside
 
     def run_command(command)
       command_name = command.join(' ')
-      run_task_response = EcsManager.run_task(config.ecs.cluster, family, command)
+      run_task_response = EcsManager.run_task(@deploy_config.cluster, family, command)
 
       unless run_task_response.successful? && run_task_response.tasks.try(:[], 0)
         exception("Failed to run #{command_name} task.", run_task_response.pretty_inspect)
@@ -243,7 +242,7 @@ module Broadside
       task_arn = run_task_response.tasks[0].task_arn
       debug "Launched #{command_name} task #{task_arn}, waiting for completion..."
 
-      EcsManager.ecs.wait_until(:tasks_stopped, { cluster: config.ecs.cluster, tasks: [task_arn] }) do |w|
+      EcsManager.ecs.wait_until(:tasks_stopped, { cluster: @deploy_config.cluster, tasks: [task_arn] }) do |w|
         w.max_attempts = nil
         w.delay = config.ecs.poll_frequency
         w.before_attempt do |attempt|
@@ -253,7 +252,7 @@ module Broadside
 
       info "#{command_name} task container logs:\n#{get_container_logs(task_arn)}"
 
-      if (code = EcsManager.get_task_exit_code(config.ecs.cluster, task_arn, family)) == 0
+      if (code = EcsManager.get_task_exit_code(@deploy_config.cluster, task_arn, family)) == 0
         debug "#{command_name} task #{task_arn} exited with status code 0"
       else
         exception "#{command_name} task #{task_arn} exited with a non-zero status code #{code}!"
@@ -261,7 +260,7 @@ module Broadside
     end
 
     def get_container_logs(task_arn)
-      ip = EcsManager.get_running_instance_ips(config.ecs.cluster, family, task_arn).first
+      ip = EcsManager.get_running_instance_ips(@deploy_config.cluster, family, task_arn).first
       debug "Found ip of container instance: #{ip}"
 
       find_container_id_cmd = "#{gen_ssh_cmd(ip)} \"docker ps -aqf 'label=com.amazonaws.ecs.task-arn=#{task_arn}'\""