+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.5.0
+- [#5](https://github.com/lumoslabs/broadside/issues/5): `:cluster` and `:poll_frequency` are now configured per target, not per broadside instance
+
 # 1.4.0
 = [#42](https://github.com/lumoslabs/broadside/pull/42/files): Update the task definition when running bootstrap